refactor: moving few constants to tft_price mod
